+++ packages/vite-plugin/src/metadata/Metadata.ts
@@ -1,7 +1,6 @@
 // SPDX-FileCopyrightText: 2023-2025 Open Pioneer project (https://github.com/open-pioneer)
 // SPDX-License-Identifier: Apache-2.0
 import { PackageConfig } from "@open-pioneer/build-common";
-import { PluginContext } from "rolldown";
 
 /**
  * Contains build-time information about an app.
@@ -49,6 +48,9 @@ export interface PackageMetadata {
 
     /** Path to package.json file. */
     packageJsonPath: string;
+
+    /** The UNRESOLVED services module id (e.g. `@foo/packageName/services`). */
+    servicesModuleId: string | undefined;
 
     /** Path to entry point (contains service exports). */
     servicesModulePath: string | undefined;
@@ -103,8 +105,6 @@ export interface PlainPackageMetadata {
  * Either an Open Pioneer Trails package or a plain package.
  */
 export type InternalPackageMetadata = PackageMetadata | PlainPackageMetadata;
-
-export type MetadataContext = Pick<PluginContext, "addWatchFile" | "resolve" | "warn">;
 
 export interface ResolvedPackageLocation {
     type: "absolute";