+++ packages/vite-plugin/src/mpaPlugin.test.ts
@@ -4,7 +4,7 @@ import { existsSync, readFileSync } from "node:fs";
 import { join, resolve } from "node:path";
 import { runViteBuild, TEMP_DATA_DIR, TEST_DATA_DIR } from "./utils/testUtils";
 import { describe, it, assert, expect } from "vitest";
-import { globSync } from "fast-glob";
+import { globSync } from "tinyglobby";
 
 describe("multi page support", function () {
     it("should include the root site if configured", async function () {