+++ packages/build-support/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog @open-pioneer/build-support
 
+## 3.0.5
+
+### Patch Changes
+
+- 2c45a18: Bump dependencies
+
+## 3.0.4
+
+### Patch Changes
+
+- f229f63: Bump dependencies
+
 ## 3.0.3
 
 ### Patch Changes