+++ packages/check-pnpm-duplicates/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog @open-pioneer/check-pnpm-duplicates
 
+## 0.2.6
+
+### Patch Changes
+
+- 2c45a18: Bump dependencies
+
+## 0.2.5
+
+### Patch Changes
+
+- f229f63: Bump dependencies
+
+## 0.2.4
+
+### Patch Changes
+
+- 1b99d82: Update eslint to version 9.
+
 ## 0.2.3
 
 ### Patch Changes